### 2018-12-13 #1127 @whedon
[PRE REVIEW]: A News Verification Browser for the Detection of Clickbait, Satire, and Falsified News
**Submitting author:** @VictoriaRubin (<a href="http://orcid.org/0000-0003-3610-9967">Victoria Rubin</a>)
**Repository:** <a href="https://github.com/litrl/litrl_code" target ="_blank">https://github.com/litrl/litrl_code</a>
**Version:** v0.12.0.0
