+++ paper.bib
@@ -9,7 +9,7 @@
 	author = {Rubin, Victoria L., Brogly, Chris, Conroy, Nadia, Chen, Yimin, Cornwell, Sarah E., Asubiaro, Toluwase V.},
 	title = {litrl/litrl_code: Litrl Browser Experimental 0.12.0.0 Public (Version exp-0.12.0.0)},
 	howpublished = {\url{https://github.com/litrl/litrl_code}},
-	year = {2018}
+	year = {2018},
 	doi = {10.5281/zenodo.2016628}
 }
 
@@ -35,7 +35,7 @@
  	acmid = {2823467},
  	publisher = {ACM},
  	address = {New York, NY, USA},
- 	keywords = {automated deception detection, clickbait, human factors, news verification, reader behavior, reliability, standardization},
+ 	keywords = {automated deception detection, clickbait, human factors, news verification, reader behavior, reliability, standardization}
 }
 
 @inproceedings{Chen:2015,
@@ -53,20 +53,21 @@
 	acmid = {2857151},
 	publisher = {American Society for Information Science},
 	address = {Silver Springs, MD, USA},
-	keywords = {credibility assessment, deception detection, fake news detection, journalism practices, media commercialization, natural language processing, news verification, online news},
+	keywords = {credibility assessment, deception detection, fake news detection, journalism practices, media commercialization, natural language processing, news verification, online news}
 } 
 
 @inproceedings{Chen:2017,
 	author = {Yimin Chen and Victoria L. Rubin},
 	title = {Perceptions of Clickbait: A Q - Methodology Approach},
 	booktitle = {Proceedings of the 45th Annual Conference of The Canadian Association for Information Science/L’Association canadienne des sciences de l’information (CAIS/ACSI2017)},
-	Series = {CAIS/ACSI 2017}
+	Series = {CAIS/ACSI 2017},
 	year = {2017},
 	location = {Ryerson University, Toronto, Canada},
 	numpages = {6},
 	URL = {https://journals.library.ualberta.ca/ojs.cais-acsi.ca/index.php/cais-asci/article/view/1046},
-	publisher = {Cameron Library, University of Alberta}
+	publisher = {Cameron Library, University of Alberta},
 	address = {Edmonton, Alberta}
+}
 
 @inproceedings{Rubin:2015,
  	author = {Rubin, Victoria L. and Chen, Yimin and Conroy, N.},
@@ -90,14 +91,14 @@
 	author = {Rubin, Victoria L.},
 	title = {News Verification Suite: Towards System Design to Supplement Reporters’ and Editors’ Judgements},
 	booktitle = {Proceedings of the 45th Annual Conference of The Canadian Association for Information Science/L’Association canadienne des sciences de l’information (CAIS/ACSI2017)},
-	Series = {CAIS/ACSI 2017}
+	Series = {CAIS/ACSI 2017},
 	year = {2017},
 	location = {Ryerson University, Toronto, Canada},
 	numpages = {6},
 	URL = {https://journals.library.ualberta.ca/ojs.cais-acsi.ca/index.php/cais-asci/article/view/1030},
-	publisher = {Cameron Library, University of Alberta}
-	address = {Edmonton, Alberta}
-	keywords = {news verification, deception detection, fake news detection, credibility assessment, reputable sources, fabrication, hoax, satire, natural language processing, text analytics, predictive modeling, corpus construction},
+	publisher = {Cameron Library, University of Alberta},
+	address = {Edmonton, Alberta},
[…]
