### 2018-12-20 69c113f
Update paper.bib
