+++ paper.bib
@@ -111,7 +111,7 @@
   	year = {2016},
   	publisher = {Association for Computational Linguistics},
   	pages = {7–17},
-	numpages = {10}
+	numpages = {10},
 	location = {San Diego, California},
   	doi = {10.18653/v1/W16-0802},
  	url = {http://aclweb.org/anthology/W16-0802}
