### 2018-12-20 ff67c1c
paper fixes
