+++ paper.bib
@@ -6,14 +6,14 @@
 }
 
 @Misc{litrl2018,
-	author = {Rubin, Victoria L., Brogly, Chris, Conroy, Nadia, Chen, Yimin, Cornwell, Sarah E., Asubiaro, Toluwase V.},
+	author = {Rubin, Victoria L. and Brogly, Chris and Conroy, Nadia and Chen, Yimin and Cornwell, Sarah E. and Asubiaro, Toluwase V.},
 	title = {litrl/litrl_code: Litrl Browser Experimental 0.12.0.0 Public (Version exp-0.12.0.0)},
 	howpublished = {\url{https://github.com/litrl/litrl_code}},
 	year = {2018},
 	doi = {10.5281/zenodo.2016628}
 }
 
-@unpublished{Brogly,
+@unpublished {Brogly,
 	author = {Chris Brogly and Victoria L. Rubin},
 	title = {Detecting Clickbait: Here’s how to Do it},
 	note = {Unpublished Manuscript},
