### 2018-12-21 #1127 @arfon
@jasonclark - would you be willing to edit this submission for JOSS?
