+++ paper.bib
@@ -1,5 +1,5 @@
 @unpublished {Asubiaro,
-	author = {Toluwase Asubiaro and Victoria L. Rubin},
+	author = {Asubiaro, Toluwase and Rubin, Victoria L.},
 	title = {Falsified News Detection Methodology},
 	note = {Unpublished Manuscript},
 	year = {2019}
@@ -41,7 +41,7 @@
 @inproceedings{Chen:2015,
 	author = {Chen, Yimin and Conroy, Niall J. and Rubin, Victoria L.},
 	title = {News in an Online World: The Need for an "Automatic Crap Detector"},
-	booktitle = {Proceedings of the 78th ASIS\&T Annual Meeting: Information Science with Impact: Research in and for the Community},
+	booktitle = {Proceedings of the 78th ASIST Annual Meeting: Information Science with Impact: Research in and for the Community},
 	series = {ASIST '15},
 	year = {2015},
 	isbn = {0-87715-547-X},
