### 2018-12-21 8983a76
Update paper.bib
