+++ paper.bib
@@ -6,7 +6,7 @@
 }
 
 @Misc{litrl2018,
-	author = {Rubin, Victoria L. and Brogly, Chris and Conroy, Nadia and Chen, Yimin and Cornwell, Sarah E. and Asubiaro, Toluwase V.},
+	author = {Rubin, Victoria L. and Brogly, Chris and Conroy, Nadia and Chen, Yimin and Cornwell, Sarah E. and Asubiaro, Toluwase},
 	title = {litrl/litrl_code: Litrl Browser Experimental 0.12.0.0 Public (Version exp-0.12.0.0)},
 	howpublished = {\url{https://github.com/litrl/litrl_code}},
 	year = {2018},
