+++ paper.bib
@@ -38,41 +38,10 @@
  	keywords = {automated deception detection, clickbait, human factors, news verification, reader behavior, reliability, standardization}
 }
 
-@inproceedings{Chen:2015,
-	author = {Chen, Yimin and Conroy, Niall J. and Rubin, Victoria L.},
-	title = {News in an Online World: The Need for an "Automatic Crap Detector"},
-	booktitle = {Proceedings of the 78th ASIST Annual Meeting: Information Science with Impact: Research in and for the Community},
-	series = {ASIST '15},
-	year = {2015},
-	isbn = {0-87715-547-X},
-	location = {St. Louis, Missouri},
-	pages = {81:1--81:4},
-	articleno = {81},
-	numpages = {4},
-	url = {http://dl.acm.org/citation.cfm?id=2857070.2857151},
-	acmid = {2857151},
-	publisher = {American Society for Information Science},
-	address = {Silver Springs, MD, USA},
-	keywords = {credibility assessment, deception detection, fake news detection, journalism practices, media commercialization, natural language processing, news verification, online news}
-} 
-
-@inproceedings{Chen:2017,
-	author = {Yimin Chen and Victoria L. Rubin},
-	title = {Perceptions of Clickbait: A Q - Methodology Approach},
-	booktitle = {Proceedings of the 45th Annual Conference of The Canadian Association for Information Science/L’Association canadienne des sciences de l’information (CAIS/ACSI2017)},
-	Series = {CAIS/ACSI 2017},
-	year = {2017},
-	location = {Ryerson University, Toronto, Canada},
-	numpages = {6},
-	URL = {https://journals.library.ualberta.ca/ojs.cais-acsi.ca/index.php/cais-asci/article/view/1046},
-	publisher = {Cameron Library, University of Alberta},
-	address = {Edmonton, Alberta}
-}
-
 @inproceedings{Rubin:2015,
  	author = {Rubin, Victoria L. and Chen, Yimin and Conroy, N.},
 	title = {Deception Detection for News: Three Types of Fakes},
-	booktitle = {Proceedings of the 78th ASIS\&T Annual Meeting: Information Science with Impact: Research in and for the Community},
+	booktitle = {Proceedings of the 78th ASIST Annual Meeting: Information Science with Impact: Research in and for the Community},
  	series = {ASIST '15},
  	year = {2015},
  	isbn = {0-87715-547-X},
