+++ paper.bib
@@ -20,10 +20,19 @@
 	year = {2019}
 }
 
+@inproceedings{Asubiaro:2018,
+	author = {Asubiaro, Toluwase and Rubin, Victoria L.},
+	title = {Comparing Features of Fabricated and Legitimate Political News in Digital Environments (2016-2017)},
+	booktitle = {Proceedings of the 81st Annual Meeting of the Association for Information Science and Technology},
+	year = {2018},
+	location = {Vancouver, Canada},
+	pages = {747-750}
+}
+
 @inproceedings{Chen:2015:MOC:2823465.2823467,
  	author = {Chen, Yimin and Conroy, Niall J. and Rubin, Victoria L.},
  	title = {Misleading Online Content: Recognizing Clickbait As "False News"},
- 	booktitle = {Proceedings of the 2015 ACM on Workshop on Multimodal Deception Detection},
+ 	booktitle = {Proceedings of the 2015 {ACM} on Workshop on Multimodal Deception Detection},
  	series = {WMDD '15},
  	year = {2015},
  	isbn = {978-1-4503-3987-2},
@@ -41,7 +50,7 @@
 @inproceedings{Rubin:2015,
  	author = {Rubin, Victoria L. and Chen, Yimin and Conroy, N.},
 	title = {Deception Detection for News: Three Types of Fakes},
-	booktitle = {Proceedings of the 78th ASIST Annual Meeting: Information Science with Impact: Research in and for the Community},
+	booktitle = {Proceedings of the 78th {ASIST} Annual Meeting: Information Science with Impact: Research in and for the Community},
  	series = {ASIST '15},
  	year = {2015},
  	isbn = {0-87715-547-X},
