### 2018-12-21 0de7350
paper edits: clarified saving results/changed ASIST references to ASIST
