+++ paper.bib
@@ -23,7 +23,7 @@
 @inproceedings{Asubiaro:2018,
 	author = {Asubiaro, Toluwase and Rubin, Victoria L.},
 	title = {Comparing Features of Fabricated and Legitimate Political News in Digital Environments (2016-2017)},
-	booktitle = {Proceedings of the 81st Annual Meeting of the Association for Information Science and Technology},
+	booktitle = {Proceedings of the 81st {ASIST} Annual Meeting},
 	year = {2018},
 	location = {Vancouver, Canada},
 	pages = {747-750}
