+++ paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 ---
 
 # Description-in-Brief
-The LiT.RL News Verification Browser is a research tool for news readers, journalists, editors or information professionals. The tool analyzes the language used in digital news web pages to determine if they are clickbait, satirical news, or falsified news, and visualizes the results by highlighting content in color-coded categories.
+The LiT.RL News Verification Browser is a research tool for news readers, journalists, editors or information professionals. The tool analyzes the language used in digital news web pages to determine if they are clickbait, satirical news, or falsified news, and visualizes the results by highlighting content in color-coded categories. All processing is completed on the local machine - clickbait, satirical news, and falsified news results are not sent to or from a remote server. Results may be saved locally to a standard SQLite database for further analysis.
 
 # Summary
 Widespread adoption of internet technologies has changed the way that news is created and consumed. The current online news environment is one that incentivizes speed and spectacle in reporting at the cost of fact-checking and verification, encouraging the proliferation of misinformation and disinformation. The LiT.RL News Verification (NV) Browser is a system that offers a first step counter-measure by automatically detecting and highlighting fake news headlines and articles. The system is presented to users as a set of assistive technologies built into a stand-alone browser tuned to identify different varieties of fakes [@Rubin:2015; @Rubin:2017]. Our algorithms look for patterns of subtle lexico-syntactic features in text. Images, audio and video formats are unsupported. The core of the functionality is in the natural language processing (NLP) of textual data and automated classification of results with machine learning using support vector machines.
