### 2018-12-21 443e6d2
paper/bib fix
added crap detector paper back
