+++ paper.bib
@@ -29,6 +29,24 @@
 	pages = {747-750}
 }
 
+@inproceedings{Chen:2015,
+	author = {Chen, Yimin and Conroy, Niall J. and Rubin, Victoria L.},
+	title = {News in an Online World: The Need for an "Automatic Crap Detector"},
+	booktitle = {Proceedings of the 78th ASIST Annual Meeting: Information Science with Impact: Research in and for the Community},
+	series = {ASIST '15},
+	year = {2015},
+	isbn = {0-87715-547-X},
+	location = {St. Louis, Missouri},
+	pages = {81:1--81:4},
+	articleno = {81},
+	numpages = {4},
+	url = {http://dl.acm.org/citation.cfm?id=2857070.2857151},
+	acmid = {2857151},
+	publisher = {American Society for Information Science},
+	address = {Silver Springs, MD, USA},
+	keywords = {credibility assessment, deception detection, fake news detection, journalism practices, media commercialization, natural language processing, news verification, online news}
+} 
+
 @inproceedings{Chen:2015:MOC:2823465.2823467,
  	author = {Chen, Yimin and Conroy, Niall J. and Rubin, Victoria L.},
  	title = {Misleading Online Content: Recognizing Clickbait As "False News"},
