+++ paper.bib
@@ -44,7 +44,8 @@
 	acmid = {2857151},
 	publisher = {American Society for Information Science},
 	address = {Silver Springs, MD, USA},
-	keywords = {credibility assessment, deception detection, fake news detection, journalism practices, media commercialization, natural language processing, news verification, online news}
+	keywords = {credibility assessment, deception detection, fake news detection, journalism practices, media commercialization, natural language processing, news verification, online news},
+	doi = {10.1002/pra2.2015.145052010081}
 } 
 
 @inproceedings{Chen:2015:MOC:2823465.2823467,
@@ -81,6 +82,7 @@
  	publisher = {American Society for Information Science},
  	address = {Silver Springs, MD, USA},
  	keywords = {corpus construction, credibility assessment, deception detection, fabrication, fake news detection, hoax, natural language processing, news verification, predictive modeling, reputable sources, satire, text analytics},
+	doi = {10.1002/pra2.2015.145052010083}
 } 
 
 @inproceedings{Rubin:2017,
