### 2018-12-21 1a95719
paper fixes
