+++ paper.bib
@@ -51,7 +51,7 @@
 @inproceedings{Chen:2015:MOC:2823465.2823467,
  	author = {Chen, Yimin and Conroy, Niall J. and Rubin, Victoria L.},
  	title = {Misleading Online Content: Recognizing Clickbait As "False News"},
- 	booktitle = {Proceedings of the 2015 {ACM} on Workshop on Multimodal Deception Detection},
+ 	booktitle = {Proceedings of the 2015 {ACM} Workshop on Multimodal Deception Detection},
  	series = {WMDD '15},
  	year = {2015},
  	isbn = {978-1-4503-3987-2},
