### 2018-12-22 e586033
Update issue templates
