### 2018-12-22 aa0ee73
Create CONTRIBUTING.md
