+++ CONTRIBUTING.md
@@ -1,0 +1,8 @@
+## Contributing
+Litrl Browser was developed by the Language and Information Technology Lab at Western University. Several RAs (Research Assistants) work in the lab. We welcome contributions from the community via pull requests. Any contributions will be reviewed by an RA and finally by the Lab Director.
+
+## Reporting issues
+Please first check the issues board to see if this problem has already been brought up. If it has not, please create a new issue.
+
+## Support
+Please first check the issues board or comment on the issues board to try and resolve any support issues. Alternatively, you can e-mail litrlbrowser@gmail.com if you have a question. As a result of lab resources, we may not be able to respond quickly or at all.
