### 2018-12-22 367bbef
Update CONTRIBUTING.md
