### 2018-12-28 #1127 @danielskatz
👋 @jasonclark - ping... (re: editing request, see https://github.com/openjournals/joss-reviews/issues/1127#issuecomment-449369691)
