### 2019-01-14 #1127 @VictoriaRubin
@arfon @danielskatz  @labarba @kyleniemeyer Request to Explain further procedures
Dear Editor-In-Chief, Associate and Associate Editors,
Could someone please explain to us how your peer-review process works?

My RA @brogly has been trying to get this simple 2 page manuscript into your peer-review pipeline for the past month, but in exactly a month nothing seems to have happened. Obviously the Holiday season is partly to blame, but there seems to be no clear path to what we are supposed to be doing next.

@danielskatz seems to imply that we need to find a editor. How do you exactly envision this? We've requested reviewers that we thought might be interested.

Is there a communication break down here? Or is our paper not conforming with any requirements to get someone to look at it please?

As much as we like the bot-enabled chat commands, I would very much appreciate some clarity from a human.

vrubin@uwo.ca
Dr. Victoria L. Rubin, Associate Professor, Faculty of Information & Media Studies (FIMS)
Director, Language & Information Technology Research Lab (LiT.RL)
FNB 4046, Western University, London, Ontario, Canada N6A 5B7
Tel: (519) 661-2111 x 88479, Fax: (519) 661-3506
