### 2019-01-14 #1127 @danielskatz
I apologize on behalf on JOSS - this has been a problem on our side, not your side.  The process is that a submission can request a specific editor, who then works with the author to find reviewers and carry out the review process.  If a submission does not request an editor, we are supposed to assign one.

In this case, our process broke: the editor who we suggested has not responded, and to some extent, blaming the holiday break, we didn't follow-up.  I will go ahead and assign myself as editor, and start the process.
