### 2019-01-14 #1127 @danielskatz
👋 @ejolly, @shivam11 - would one (or ideally both) of you be willing to review this for JOSS?
