### 2019-01-14 #1127 @VictoriaRubin
@danielskatz - Thanks for clarifying. Re outside reviewers: perhaps, Martin Potthast, who ran the CLickbait Challenge (https://webis.de/data/webis-clickbait-17.html)? Or anyone from his team or the other dataset group?
@brogly  Chris, please stay on this, to make sure no further glitches occur. You had Martin Potthast's e-mail somewhere when we asked for permissions to use his set, right? Just in case @danielskatz asks for it Thanks again to both of you!

See 2 references below from our GitHub:
[CLICKBAIT DATASET 1] Abhijnan Chakraborty, Bhargavi Paranjape, Sourya Kakarla, and Niloy Ganguly. "Stop Clickbait: Detecting and Preventing Clickbaits in Online News Media”. In Proceedings of the 2016 IEEE/ACM International Conference on Advances in Social Networks Analysis and Mining (ASONAM), San Fransisco, US, August 2016. (URL: https://github.com/bhargaviparanjape/clickbait)

[CLICKBAIT DATASET 2] Martin Potthast, Tim Gollub, Kristof Komlossy, Sebastian Schuster, Matti Wiegmann, Erika Patricia Garces Fernandez, Matthias Hagen, and Benno Stein. Crowdsourcing a Large Corpus of Clickbait on Twitter. In Proceedings of the 27th International Conference on Computational Linguistics (COLING 2018), pages 1498–1507, August 2018. The COLING 2018 Organizing Committee.
