### 2019-01-14 #1127 @ejolly
> 👋 @ejolly, @shivam11 - would one (or ideally both) of you be willing to review this for JOSS?

Unfortunately, I don't have easy access to a Windows computer (OSX and Linux only), so I'm not sure I'll be able to review this submission. Looks really interesting though!
