### 2019-01-15 #1127 @danielskatz
👋 @CBenghi, @xirdneh, @nmullins, @dredamonsta1 - would one of you be interested in being the second reviewer for this submission on JOSS?
