### 2019-01-15 #1127 @danielskatz
👋 @bmst, @johanneskiesel, @kleinfreund, @phoerious - would one you be willing to review this submission for JOSS? Or perhaps you can suggest someone from your group/team? (See the [JOSS review guidelines](
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) in case you are not familiar with how JOSS works)
