### 2019-01-15 #1127 @potthast
Thanks for the invitation.

Apart from me, only @bmst has been doing research on clickbait detection lately. Unfortunately, at present, we are both overwhelmed with other commitments, so that we will have to pass.

Sorry for the inconvenience!
