### 2019-01-17 #1127 @danielskatz
👋 @felixhenninger or @jaynanavati - would you be willing to review this for JOSS?
