### 2019-01-18 #1127 @danielskatz
👋 @theBrianCui or @rahulkapoor90 - would one of you be willing to review this for JOSS?  (See the [JOSS review guidelines](
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) in case you are not familiar with how JOSS works)
