### 2019-01-21 #1127 @brogly
@lrasmus - Thanks for agreeing to review.
@theBrianCui - We are looking for a second reviewer for the Litrl Browser software and associated paper for submission to JOSS. Links to these are above and @danielskatz (Editor) posted review guidelines right above your reply if you are interested.
@danielskatz - The following reviewers may also be interested, chkoar, sirsharpest, rljacobson, lindemann09, adammichaelwood (omitting the @ as I'm not sure how many people I'm supposed to ask all at once...)
