### 2019-01-22 #1127 @danielskatz
Hi @theBrianCui - sorry for my delay in responding.  JOSS is an open journal for short articles that also contain open source research software.  Review is of both the article and the software.  Once a submission is reviewed successfully (which means the reviewers point out problems and the authors fix them), the submission is published and the authors can then use this publication for academic credit (by asking users to cite the publication).  [Here](
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) are the review guidelines.

Are you willing to review this submission? (this issue overall, and [this article draft](
https://github.com/openjournals/joss-papers/blob/joss.01127/joss.01127/10.21105.joss.01127.pdf) in particular)?
