### 2019-01-27 #1208 @whedon
[REVIEW]: A News Verification Browser for the Detection of Clickbait, Satire, and Falsified News
**Submitting author:** @VictoriaRubin (<a href="http://orcid.org/0000-0003-3610-9967">Victoria Rubin</a>)
**Repository:** <a href="https://github.com/litrl/litrl_code" target ="_blank">https://github.com/litrl/litrl_code</a>
**Version:** v0.14.0.0

- [x] **Version:** v0.14.0.0

- [x] **Version:** v0.14.0.0
