+++ paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 ---
 
 # Description-in-Brief
-The LiT.RL News Verification Browser is a research tool for news readers, journalists, editors or information professionals. The tool analyzes the language used in digital news web pages to determine if they are clickbait, satirical news, or falsified news, and visualizes the results by highlighting content in color-coded categories. All processing is completed on the local machine - clickbait, satirical news, and falsified news results are not sent to or from a remote server. Results may be saved locally to a standard SQLite database for further analysis.
+The LiT.RL News Verification Browser is a research tool for news readers, journalists, editors or information professionals. The tool analyzes the language used in digital news web pages to determine if they are clickbait, satirical news, or falsified news, and visualizes the results by highlighting content in color-coded categories. Although the clickbait, satire, and falsification detectors perform to certain accuracy levels on test data, during real-world internet use accuracy may vary. The browser is not a replacement for digital literacy and is not always correct. All processing is completed on the local machine - results are not sent to or from a remote server. Results may be saved locally to a standard SQLite database for further analysis.
 
 # Summary
 Widespread adoption of internet technologies has changed the way that news is created and consumed. The current online news environment is one that incentivizes speed and spectacle in reporting at the cost of fact-checking and verification, encouraging the proliferation of misinformation and disinformation. The LiT.RL News Verification (NV) Browser is a system that offers a first step counter-measure by automatically detecting and highlighting fake news headlines and articles. The system is presented to users as a set of assistive technologies built into a stand-alone browser tuned to identify different varieties of fakes [@Rubin:2015; @Rubin:2017]. Our algorithms look for patterns of subtle lexico-syntactic features in text. Images, audio and video formats are unsupported. The core of the functionality is in the natural language processing (NLP) of textual data and automated classification of results with machine learning using support vector machines.
@@ -52,7 +52,7 @@ Each is implemented as a separate overlay option through three tabs at the botto
 ![Figure 1: Screenshot of LiT.RL News Verification Browser clickbait detection on the CNN homepage (December 12, 2018)](figure.png)
 
 # Limitations
-This browser is meant to augment our human discernment, rather than replace it, by highlighting potentially false information which may require further scrutiny. Digital literacy is key for everyone to effectively evaluate potential misinformation online, and the LiT.RL Browser is NOT a replacement for that [@Chen:2015]. News readers’ critical thinking remains key to navigating the increasingly fraught online information landscape. The News Verification Browser is a research tool - not a replacement for a day-to-day web browser. It should not be used where security is critical. 
+This browser is meant to augment our human discernment, rather than replace it, by highlighting potentially false information which may require further scrutiny. Digital literacy is key for everyone to effectively evaluate potential misinformation online, and the LiT.RL Browser is NOT a replacement for that [@Chen:2015]. During real-world internet use, the accuracy of the detectors in this tool may vary. News readers’ critical thinking remains key to navigating the increasingly fraught online information landscape. The News Verification Browser is a research tool - not a replacement for a day-to-day web browser. It should not be used where security is critical. 
 
 # Availability
  The source code is openly available on GitHub [@litrl2018] under the GPLv3 license for anyone in the research and development community to use or improve on. The public can download the browser for experimentation on their own computers with “no strings attached”. 
