### 2019-02-05 #1208 @lrasmus
> 👋 @lrasmus @CBenghi - Can I get an update from you on this review?

@danielskatz I have started with the basic checklist items, sorry it was a little slow start on my end.  I am planning to complete by the end of this week.  Thank you @brogly for updates in manuscript and instructions.
