### 2019-02-05 #1208 @lrasmus
@danielskatz - clarification question.  The authors have appropriately tagged and marked for release v0.12.0.0 of the software.  Should we review just what is in the repository as of that tag, or include recent commits?  Not trying to be a stickler, just wondering if recommended practice is for authors to re-tag software updates that we then review, or if software gets versioned and tagged following completion of JOSS review.  Thanks!
