### 2019-02-05 #1208 @danielskatz
You should review changes, particularly those made in response to review issues.

>software gets versioned and tagged following completion of JOSS review

yes, this is what happens
