+++ paper.bib
@@ -1,7 +1,7 @@
-@unpublished {Asubiaro,
+@article {Asubiaro,
 	author = {Asubiaro, Toluwase and Rubin, Victoria L.},
 	title = {Falsified News Detection Methodology},
-	note = {Unpublished Manuscript},
+	journal = {Unpublished article},
 	year = {2019}
 }
 
@@ -13,10 +13,10 @@
 	doi = {10.5281/zenodo.2016628}
 }
 
-@unpublished {Brogly,
+@article {Brogly,
 	author = {Chris Brogly and Victoria L. Rubin},
 	title = {Detecting Clickbait: Here’s how to Do it},
-	note = {Unpublished Manuscript},
+	journal = {Unpublished article under review},
 	year = {2019}
 }
 
