### 2019-02-11 #1208 @brogly
@lrasmus @CBenghi minor change to PDF references to clarify that Asubiaro & Rubin and Brogly & Rubin are both unpublished articles
