### 2019-02-12 litrl/litrl_code#19 @CBenghi
[FEATURE REQUEST] Upgrade to Python 3
**Describe the issue**
Executing the installation command we get notified that Python 2.7 is reaching end of life support in January 2020, less than an year.

"DEPRECATION: Python 2.7 will reach the end of its life on January 1st, 2020. Please upgrade your Python as Python 2.7 won't be maintained after that date. A future version of pip will drop support for Python 2.7."
Pip also notified that version 2.7 is not going to be maintained.

**Expected behavior**
Research software that can be usefully picked up by other teams should be reliable for longer than that timespan.

**Additional context**
Would it be complex to update the installation to a current version of Python, with longer support?
