### 2019-02-12 litrl/litrl_code#19 @CBenghi
Please note that I did not mean to mark this as a bug, but that was done by the template.
I'm not sure how to remove the "bug" label.
