### 2019-02-12 litrl/litrl_code#21 @CBenghi
[BUG] Form ratio on high-dpi monitors
**Describe the bug**
When using the program in an high-dpi monitor some settings are not readable.

**To Reproduce**
Steps to reproduce the behavior:
1. Change the dpi settings in windows 10 for instance to to 200%
2. Click on 'Highlight options' button on the right hand side of the page.
3. See truncated text.

**Expected behavior**
All text should be readable, or have a tooltip help at least.

**Screenshots**
![image](https://user-images.githubusercontent.com/5476273/52652929-43bc8300-2eef-11e9-9511-3a982f2aabdb.png)

 - OS: [Win10]
 - Version [0.12.0.0]
 - Last URL: Irrelevant
