### 2019-02-12 litrl/litrl_code#19 @brogly-zz
@CBenghi - Thanks for bringing this up - I was going to add this myself at some point - I have added this to the next milestone. It is unclear to me how much effort this would require; it could be minimal or it may take a while to move everything over. In any case, we will work to complete this before 2.7 is outdated. If you think it should be done for JOSS, I can work on that but our capacity for significant changes is limited right now.
