### 2019-02-12 litrl/litrl_code#19 @CBenghi
@brogly,
In principle I've added it w/reg Joss, but I'm not sure about the policy for this.
I appreciate that it could be non trivial, but I'm not much familiar with python.

I'll have to speak to the editor about that.
