### 2019-02-15 #1208 @lrasmus
@danielskatz - making progress.  @CBenghi is rocking it reporting technical issues/considerations (https://github.com/litrl/litrl_code/issues).  I am finding some documentation gaps that I am opening issues for in the repo as well.
