+++ paper.bib
@@ -26,7 +26,8 @@
 	booktitle = {Proceedings of the 81st {ASIST} Annual Meeting},
 	year = {2018},
 	location = {Vancouver, Canada},
-	pages = {747-750}
+	pages = {747-750},
+	doi = {10.1002/pra2.2018.14505501100}
 }
 
 @inproceedings{Chen:2015,
