### 2019-02-16 #1208 @brogly
@CBenghi @lrasmus - Added a newly assigned DOI for one of the references
