### 2019-02-16 86bf940
add clickbait test doc and fix for some highlight issues
some highlighting issues caused by detectors ignoring unicode special chars related to quotations etc, easiest fix is to ignore these when highlighting links until move to python 3 happens
