+++ vs_solution/LITRL/Decoy/Properties/AssemblyInfo.cs
@@ -11,7 +11,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("FIMS, Western University")]
 [assembly: AssemblyProduct("Litrl Browser EXPERIMENTAL")]
-[assembly: AssemblyCopyright("Copyright Victoria L Rubin 2017-2018")]
+[assembly: AssemblyCopyright("Copyright Victoria L Rubin 2017-2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -33,7 +33,7 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.12.0.0")]
-[assembly: AssemblyFileVersion("0.12.0.0")]
+[assembly: AssemblyVersion("0.14.0.0")]
+[assembly: AssemblyFileVersion("0.14.0.0")]
 [assembly: NeutralResourcesLanguage("en")]
 
