### 2019-02-16 639d1d3
improve link highlighting/update warning message
