### 2019-02-19 litrl/litrl_code#21 @brogly-zz
@CBenghi - I have duplicated the issue and am working on a fix. It might take a few more days but I'm hoping to resolve this shortly.
