### 2019-02-21 #1208 @brogly
@lrasmus @CBenghi - Updated document to clarify detector scores as requested by @lrasmus
