### 2019-02-21 #1208 @lrasmus
@danielskatz , @brogly, @CBenghi - I've completed my review and feel this meets the JOSS review guidelines.
