### 2019-02-22 #1208 @CBenghi
@danielskatz,

I've given it another look yesterday and I've seen a lot of progress.
I was waiting for @brogly to complete https://github.com/litrl/litrl_code/issues/21 as he last stated a few days ago to give another look.

@danielskatz, perhaps you can also help clear a doubt that I've expressed at https://github.com/litrl/litrl_code/issues/19, the issue is around the long term reliability of the software as the version of python that it uses reaches end of life in less than a year. I'm not sure what is the policy that Joss adopts around this.
