### 2019-02-22 #1208 @danielskatz
Thanks.

> @danielskatz, perhaps you can also help clear a doubt that I've expressed at [litrl/litrl_code#19](https://github.com/litrl/litrl_code/issues/19), the issue is around the long term reliability of the software as the version of python that it uses reaches end of life in less than a year. I'm not sure what is the policy that Joss adopts around this.

In this case, given the statement in https://github.com/litrl/litrl_code/issues/19 that this will be fixed before it is an issue, I don't think it should block acceptance of this version of the software.
