### 2019-02-22 #1208 @lrasmus
I should echo that I think @CBenghi raised some really good issues.  This was my first JOSS review, and I wasn't sure if any of those would be blockers for acceptance or not.
