### 2019-02-22 #1208 @brogly
@danielskatz - Thanks - we are happy to upgrade the software to Python 3 before 2020 but basically the problem is doing it *right now* - we will have more time for that fix later in the year.

@lrasmus - Thanks for the excellent feedback throughout.

@CBenghi - Still working on the DPI issue and hope to get that fixed today.
