+++ vs_solution/LITRL/Decoy/FrmColorSettings.Designer.cs
@@ -89,20 +89,20 @@
             this.groupBox1.Controls.Add(this.label2);
             this.groupBox1.Controls.Add(this.label1);
             this.groupBox1.Controls.Add(this.panSlightCB);
-            this.groupBox1.Location = new System.Drawing.Point(13, 12);
+            this.groupBox1.Location = new System.Drawing.Point(12, 11);
             this.groupBox1.Margin = new System.Windows.Forms.Padding(4);
             this.groupBox1.Name = "groupBox1";
             this.groupBox1.Padding = new System.Windows.Forms.Padding(4);
-            this.groupBox1.Size = new System.Drawing.Size(471, 243);
+            this.groupBox1.Size = new System.Drawing.Size(565, 236);
             this.groupBox1.TabIndex = 0;
             this.groupBox1.TabStop = false;
             this.groupBox1.Text = "Clickbait";
             // 
             // BtnSetHeavyCB
             // 
-            this.BtnSetHeavyCB.Location = new System.Drawing.Point(322, 183);
+            this.BtnSetHeavyCB.Location = new System.Drawing.Point(329, 165);
             this.BtnSetHeavyCB.Name = "BtnSetHeavyCB";
-            this.BtnSetHeavyCB.Size = new System.Drawing.Size(64, 28);
+            this.BtnSetHeavyCB.Size = new System.Drawing.Size(87, 37);
             this.BtnSetHeavyCB.TabIndex = 14;
             this.BtnSetHeavyCB.Text = "Set";
             this.BtnSetHeavyCB.UseVisualStyleBackColor = true;
@@ -110,9 +110,9 @@
             // 
             // panSetModerateCB
             // 
-            this.panSetModerateCB.Location = new System.Drawing.Point(322, 84);
+            this.panSetModerateCB.Location = new System.Drawing.Point(329, 75);
             this.panSetModerateCB.Name = "panSetModerateCB";
-            this.panSetModerateCB.Size = new System.Drawing.Size(64, 29);
+            this.panSetModerateCB.Size = new System.Drawing.Size(87, 39);
             this.panSetModerateCB.TabIndex = 13;
             this.panSetModerateCB.Text = "Set";
             this.panSetModerateCB.UseVisualStyleBackColor = true;
@@ -120,9 +120,9 @@
             // 
             // BtnSetNotCB
             // 
-            this.BtnSetNotCB.Location = new System.Drawing.Point(104, 84);
+            this.BtnSetNotCB.Location = new System.Drawing.Point(79, 75);
             this.BtnSetNotCB.Name = "BtnSetNotCB";
-            this.BtnSetNotCB.Size = new System.Drawing.Size(64, 29);
+            this.BtnSetNotCB.Size = new System.Drawing.Size(80, 39);
             this.BtnSetNotCB.TabIndex = 12;
             this.BtnSetNotCB.Text = "Set";
             this.BtnSetNotCB.UseVisualStyleBackColor = true;
@@ -130,9 +130,9 @@
             // 
             // BtnSetSlightCB
             // 
-            this.BtnSetSlightCB.Location = new System.Drawing.Point(104, 183);
+            this.BtnSetSlightCB.Location = new System.Drawing.Point(79, 165);
             this.BtnSetSlightCB.Name = "BtnSetSlightCB";
-            this.BtnSetSlightCB.Size = new System.Drawing.Size(64, 28);
+            this.BtnSetSlightCB.Size = new System.Drawing.Size(80, 37);
             this.BtnSetSlightCB.TabIndex = 11;
             this.BtnSetSlightCB.Text = "Set";
             this.BtnSetSlightCB.UseVisualStyleBackColor = true;
@@ -141,77 +141,77 @@
[…]
