+++ vs_solution/LITRL/Decoy/FrmMain.Designer.cs
@@ -196,8 +196,8 @@
             // 
             this.splitContainer1.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.splitContainer1.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.splitContainer1.Location = new System.Drawing.Point(0, 38);
-            this.splitContainer1.Margin = new System.Windows.Forms.Padding(2, 3, 2, 3);
+            this.splitContainer1.Location = new System.Drawing.Point(0, 71);
+            this.splitContainer1.Margin = new System.Windows.Forms.Padding(4, 5, 4, 5);
             this.splitContainer1.Name = "splitContainer1";
             // 
             // splitContainer1.Panel1
@@ -210,8 +210,9 @@
             this.splitContainer1.Panel2.BackColor = System.Drawing.SystemColors.ControlLightLight;
             this.splitContainer1.Panel2.Controls.Add(this.TabsDetectors);
             this.splitContainer1.Panel2MinSize = 510;
-            this.splitContainer1.Size = new System.Drawing.Size(1184, 640);
-            this.splitContainer1.SplitterDistance = 669;
+            this.splitContainer1.Size = new System.Drawing.Size(2368, 968);
+            this.splitContainer1.SplitterDistance = 1337;
+            this.splitContainer1.SplitterWidth = 8;
             this.splitContainer1.TabIndex = 4;
             // 
             // TabsBrowser
@@ -223,21 +224,21 @@
             this.TabsBrowser.Dock = System.Windows.Forms.DockStyle.Fill;
             this.TabsBrowser.Font = new System.Drawing.Font("Arial", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.TabsBrowser.Location = new System.Drawing.Point(0, 0);
-            this.TabsBrowser.Margin = new System.Windows.Forms.Padding(2, 3, 2, 3);
+            this.TabsBrowser.Margin = new System.Windows.Forms.Padding(4, 5, 4, 5);
             this.TabsBrowser.Name = "TabsBrowser";
             this.TabsBrowser.SelectedIndex = 0;
-            this.TabsBrowser.Size = new System.Drawing.Size(667, 638);
+            this.TabsBrowser.Size = new System.Drawing.Size(1335, 966);
             this.TabsBrowser.TabIndex = 0;
             this.TabsBrowser.SelectedIndexChanged += new System.EventHandler(this.TabsBrowser_SelectedIndexChanged);
             this.TabsBrowser.Selected += new System.Windows.Forms.TabControlEventHandler(this.TabsBrowser_Selected);
             // 
             // TabDocument
             // 
-            this.TabDocument.Location = new System.Drawing.Point(4, 4);
-            this.TabDocument.Margin = new System.Windows.Forms.Padding(2, 3, 2, 3);
+            this.TabDocument.Location = new System.Drawing.Point(8, 8);
+            this.TabDocument.Margin = new System.Windows.Forms.Padding(4, 5, 4, 5);
             this.TabDocument.Name = "TabDocument";
-            this.TabDocument.Padding = new System.Windows.Forms.Padding(2, 3, 2, 3);
-            this.TabDocument.Size = new System.Drawing.Size(659, 607);
+            this.TabDocument.Padding = new System.Windows.Forms.Padding(4, 5, 4, 5);
+            this.TabDocument.Size = new System.Drawing.Size(1319, 908);
             this.TabDocument.TabIndex = 0;
             this.TabDocument.Text = "BROWSER";
[…]
