### 2019-02-22 litrl/litrl_code#21 @brogly-zz
@CBenghi - I've done my best to address this issue with some changes. The downside is that the form looks a little odd on normal DPI screens but it is usable up to 200% now.
