+++ vs_solution/LITRL/Decoy/FrmMain.Designer.cs
@@ -92,7 +92,6 @@
             this.BtnTagSelectClickbait = new System.Windows.Forms.Button();
             this.NumMinWordCount = new System.Windows.Forms.NumericUpDown();
             this.label2 = new System.Windows.Forms.Label();
-            this.zedGraphClickbait = new ZedGraph.ZedGraphControl();
             this.txtBottomLine = new System.Windows.Forms.TextBox();
             this.tabPage2 = new System.Windows.Forms.TabPage();
             this.splitContainer2 = new System.Windows.Forms.SplitContainer();
@@ -148,6 +147,7 @@
             this.BtnForward = new System.Windows.Forms.Button();
             this.BtnBack = new System.Windows.Forms.Button();
             this.toolTip1 = new System.Windows.Forms.ToolTip(this.components);
+            this.zedGraphClickbait = new ZedGraph.ZedGraphControl();
             ((System.ComponentModel.ISupportInitialize)(this.splitContainer1)).BeginInit();
             this.splitContainer1.Panel1.SuspendLayout();
             this.splitContainer1.Panel2.SuspendLayout();
@@ -196,8 +196,8 @@
             // 
             this.splitContainer1.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
             this.splitContainer1.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.splitContainer1.Location = new System.Drawing.Point(0, 71);
-            this.splitContainer1.Margin = new System.Windows.Forms.Padding(4, 5, 4, 5);
+            this.splitContainer1.Location = new System.Drawing.Point(0, 37);
+            this.splitContainer1.Margin = new System.Windows.Forms.Padding(2, 3, 2, 3);
             this.splitContainer1.Name = "splitContainer1";
             // 
             // splitContainer1.Panel1
@@ -210,9 +210,8 @@
             this.splitContainer1.Panel2.BackColor = System.Drawing.SystemColors.ControlLightLight;
             this.splitContainer1.Panel2.Controls.Add(this.TabsDetectors);
             this.splitContainer1.Panel2MinSize = 510;
-            this.splitContainer1.Size = new System.Drawing.Size(2368, 968);
-            this.splitContainer1.SplitterDistance = 1337;
-            this.splitContainer1.SplitterWidth = 8;
+            this.splitContainer1.Size = new System.Drawing.Size(1148, 522);
+            this.splitContainer1.SplitterDistance = 634;
             this.splitContainer1.TabIndex = 4;
             // 
             // TabsBrowser
@@ -224,21 +223,21 @@
             this.TabsBrowser.Dock = System.Windows.Forms.DockStyle.Fill;
             this.TabsBrowser.Font = new System.Drawing.Font("Arial", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.TabsBrowser.Location = new System.Drawing.Point(0, 0);
-            this.TabsBrowser.Margin = new System.Windows.Forms.Padding(4, 5, 4, 5);
+            this.TabsBrowser.Margin = new System.Windows.Forms.Padding(2, 3, 2, 3);
             this.TabsBrowser.Name = "TabsBrowser";
             this.TabsBrowser.SelectedIndex = 0;
-            this.TabsBrowser.Size = new System.Drawing.Size(1335, 966);
+            this.TabsBrowser.Size = new System.Drawing.Size(632, 520);
             this.TabsBrowser.TabIndex = 0;
             this.TabsBrowser.SelectedIndexChanged += new System.EventHandler(this.TabsBrowser_SelectedIndexChanged);
[…]
