### 2019-02-24 litrl/litrl_code#21 @CBenghi
thanks @brogly, I'll review tomorrow.
I'm in a busy time right now, but I'll do my best to complete the whole review in a couple of days.
Claudio
