### 2019-03-05 #1208 @CBenghi
Hi All,
I've been able to spend a bit more time and I feel we are almost there, but I have a couple of minor issues.

@danielskatz version of the software should be updated to v0.14.0.0, @whedon rejected my command there.

@brogly, two outstanding aspects to look at for me:
1) I'm uncertain on the benefit or referencing unpublished work (lacking DOI as per checklist above), I would probably remove them at this stage.
2) I've also opened an issue at: https://github.com/litrl/litrl_code/issues/27, I've provided two easy courses of action there, in one of them you'd need to setup a test harness, which I would be happy to help with.

Once these are closed I'm happy to accept.

Best,
Claudio
