### 2019-03-05 #1208 @danielskatz
@CBenghi - referencing unpublished work is ok - the checkbox just says that for every reference that has a DOI, the DOI should be listed.  If a reference doesn't have a DOI, that's fine.
