### 2019-03-08 litrl/litrl_code#27 @brogly-zz
@CBenghi - I have split the clickbait tests file into "clickbait_tests.html" and "clickbait_tests_for_improvements.html". The second file contains misclassified headlines, or known fails, that we can use to improve the detector. I've also added an "ACCEPTANCE CRITERIA" label to the satire and the falsification tests to clarify what colour they should be (by default).

I like the threshold idea but I think we would need to collect more headlines for the categories.

Let me know if this is sufficient - thanks for pointing this out!
