### 2019-03-09 litrl/litrl_code#27 @CBenghi
Hi @brogly,
I'll accept this solution for the paper.
Conceptually  I'd like to keep it open in this repo, becasue in the long term I think the threshold idea would be a better way to test the application.

But feel free to close if you think otherwise.

Best,
Claudio
