### 2019-03-09 #1208 @CBenghi
Dear @arfon,

I'm glad to let you know that @brogly has addressed all the issues.
I agree with @lrasmus that the repository passes all requirements for the review and I've flagged them all above.

Well done @brogly, I'm tempted to start using this myself in normal web browsing :-)

Best,
Claudio
