### 2019-03-09 litrl/litrl_code#27 @brogly-zz
@CBenghi - Thanks - and I will leave this open because this would be a nice approach to testing things. I'll see if some of the other Research Assistants can assist with more texts.
